Move the job of creating the fileshare symlink from xdg-autostart to user-mode systemd
This has the advantage of making the symlink available if a user logs
in via the command line instead of using the xfce desktop.  There have
also been some issues with the existing symlink (see issue #17) and
this change should resolve that since (1) the symlink is being created
by the user instead of root, and (2) with this method the symlink will
be created _after_ /share is mounted.

Also update the molecule tests accordingly.