Use systemd-enabled Docker images for this Ansible role
Since we are copying a systemd service file to /etc/systemd/user, we
need that directory to exist.  The best way to ensure this is to
ensure that systemd is installed.

Note that this also requires updating the prepare playbook to include
the code that waits until systemd has fully initialized on Fedora
systems before attempting to install any packages.