Make the enabling of the systemd user-mode service work outside of Molecule
The task that actually enables the user-mode systemd service can
succeed when run via Molecule, but as far as I can tell the only way
to make this work is to set "become" to "no" and set
"ansible_ssh_user" to "{{ username }}".  Unfortunately this
configuration cannot work when this role is used to build an AMI,
since in that case ssh connects via an AWS-generated ssh key that is
only valid for the AMI's default user.  Therefore we take the even
more drastic measure of applying the molecule-notest tag to this task.

I created #29 as a reminder to figure out a way to make the
ansible.builtin.systemd task succeed when run via molecule or when
this Ansible role is used to built an AMI.