Use the sudosu become method when enabling the user's systemd user service
Using su (together with sudo, since we are not sshing in as the root
user) to become the user causes an actual systemd session to be
created for the user.  This in turn causes the user's systemd user
services to be started, which we require for our Molecule tests.

Note that we also install the Python ansible package in our pre-commit
configuration.  This is necessary because the community.general.sudosu
become method is packaged there instead of in the ansible-core Python
package.